chore: add documentation for the build scripts
