+++ .travis.yml
@@ -16,7 +16,7 @@ cache:
 install:
   - npm ci
 script:
-  - npm run build:data-fetch
+  - npm run scripts:build:data
   - npm run build
   - npm run cy:serve &
   - npm run cy:run
